Viewer refactor - add some directory hierarchy
