+++ src/viewer/assets/lib/charts/util/colormap.js
