+++ src/viewer/assets/lib/charts/util/utils.js
@@ -1,0 +1,9 @@
+/**
+ * Format a timestamp for display
+ * @param {number} timestamp - Unix timestamp in milliseconds
+ * @returns {string} Formatted date/time string
+ */
+export function formatDateTime(timestamp) {
+    const date = new Date(timestamp);
+    return `${date.toISOString().slice(0, 10)} ${String(date.getHours()).padStart(2, '0')}:${String(date.getMinutes()).padStart(2, '0')}:${String(date.getSeconds()).padStart(2, '0')}`;
+}
